requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pycountry
replicate
python-multipart
google-generativeai
pillow